Renomeando e melhorando Pessoa para Paciente
