+++ src/main/java/com/connect/jpa/repository/PessoaFisicaRepository.java
@@ -3,9 +3,9 @@ package com.connect.jpa.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.connect.jpa.model.PessoaFisicaModel;
+import com.connect.jpa.model.PacienteModel;
 
 @Repository
-public interface PessoaFisicaRepository extends JpaRepository<PessoaFisicaModel, Long> {
+public interface PessoaFisicaRepository extends JpaRepository<PacienteModel, Long> {
 	
 }
