+++ src/main/java/com/connect/jpa/repository/UserInfoRepository.java
@@ -3,11 +3,11 @@ package com.connect.jpa.repository;
 import org.springframework.data.jpa.repository.JpaRepository; 
 import org.springframework.stereotype.Repository;
 
-import com.connect.jpa.model.UserInfo;
+import com.connect.jpa.model.Usuario;
 
 import java.util.Optional; 
   
 @Repository
-public interface UserInfoRepository extends JpaRepository<UserInfo, Long> { 
-    Optional<UserInfo> findByNome(String username); 
+public interface UserInfoRepository extends JpaRepository<Usuario, Long> { 
+    Optional<Usuario> findByNome(String username); 
 }
